move twittersession call to ListsActivity
