+++ app/src/main/res/layout/activity_lists.xml
@@ -9,11 +9,5 @@
         android:gravity="center_horizontal|center_vertical"
         android:text="No Tweets"/>
 
-    <ListView android:id="@id/android:list"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="1"
-        android:divider="#e1e8ed"
-        android:dividerHeight="1dp"
-        android:drawSelectorOnTop="false"/>
+
 </LinearLayout>
